Refactoring launching Vanilla.
